fix: resolve CI failures in TS lint and Rust dylint for team command
Move oldteam to correct alphabetical position in cspell.json.
Relocate inline test module to external team/tests.rs file.
Replace wildcard import with explicit imports in tests.
Use raw string literals instead of escaped quotes in JSON assertions.
Add trailing commas to multiline macro invocations.
Rename single-letter closure params to descriptive names.
Remove unnecessary allow(clippy::needless_lifetimes).
