+++ cspell.json
@@ -218,7 +218,6 @@
     "nerfed",
     "newteam",
     "newversion",
-    "oldteam",
     "nistp",
     "NOASSERTION",
     "nodetouch",
@@ -239,6 +238,7 @@
     "ntfs",
     "nushell",
     "ofjergrg",
+    "oldteam",
     "onclickoutside",
     "oomol",
     "openharmony",
